builtin/diff: implement batch blob diff option
When generating a diff between a pair of blobs, users may want the
option to perform this operation in batches by providing input on stdin.
To facilitate this, the "--batch-blobs" option is added to git-diff(1).

When provided, each line of stdin is parsed for two revisions that each
resolve to a blob. This blob pair then queues and flushes a diff to
output. This is repeated until there is no longer input on stdin.

Signed-off-by: Justin Tobler <jltobler@gmail.com>