builtin: introduce diff-blob command
Through git-diff(1), a single diff can be generated from a pair of blob
revisions directly. This is useful when more explicit control over the
set of diffs to compute is desired. Expanding on this, it would be
useful to also support generating multiple diffs for blob pairs provided
on stdin to faciliate batch processing in a single process. Batch blob
diff processing is likely considered more of a plumbing feature so,
instead of further extending the porcelain git-diff(1), a diff plumbing
command should be used.

As there is not an existing diff plumbing command that handles blob
diffs, introduce git-diff-blob(1) which generates a single diff between
a specified pair of blobs following how the same operation is done in
git-diff(1). While git-diff-blob(1) functionality is a direct subset of
git-diff(1), a subsequent patch extends it to provide a new plumbing
related feature.

The surrounding setup required for the new builtin is also added. An
existing test for blob diffs through git-diff(1) is also modified to
reuse its test cases for git-diff-blob(1).

Signed-off-by: Justin Tobler <jltobler@gmail.com>