builtin/diff-blob: add "--stdin" option
There is not a way to generate multiple blob diffs from a single
process. Similar to git-diff-tree(1) with its "--stdin" option, it would
be useful if multiple blob pairs could be provided to git-diff-blob(1)
to compute blob diffs for.

Teach git-diff-blob(1) the "--stdin" option to allow a pair of blobs to
be read from each line of stdin instead of relying on the single blob
pair provided as arguments. When this option is specified, each valid
line of input computes a blob diff thus allowing multiple blob diffs in
a single process. A blob may be specified by its ID or a path-scoped
revision that resolve to a blob. When a path-scoped revision is used,
path and mode information is also extracted and presented in the
resulting diff header.

Signed-off-by: Justin Tobler <jltobler@gmail.com>