builtin/diff-blob: Add "-z" option
The "--stdin" option for git-diff-blob(1) reads two space separated
blobs for each line of input. A blob may be specified by its ID or a
path-scoped revision that resolves to a blob. It is possible for the
path to contain whitespace or newline characters which must be escaped.

To make input more simple, teach git-diff-blob(1) the "-z" option which
changes the input delimiter for each blob to a NUL character. With this
option, the command waits two NUL terminated blobs to read and then
generates the diff. The diff output is also NUL terminated to help
differentiate between outputted diffs.

Signed-off-by: Justin Tobler <jltobler@gmail.com>