format-patch: add option to generate CC list
In git-format-patch(1), the `--cc` option appends a specified email
address to be CC'd on all generated patches. Additionally, via the
`format.cc` configuration, CC recipients can be preconfigured. However,
if a user wants to generate the CC list programmatically, there is no
option to do so.

Introduce the `--cc-cmd` flag and `format.ccCmd` configuration to
optionally specify a program to execute. Its output is interpreted as a
list of additional email addresses to CC on all patches. Each line of
stdout is expected to contain a single email address.

Signed-off-by: Justin Tobler <jltobler@gmail.com>