builtin/update-index: skip transaction with stdin and verbose options
With 23a3a303ab9 (update-index: use the bulk-checkin infrastructure,
2022-04-04), object database transactions were added to
git-update-index(1) to facilitate writing objects in bulk. With
transactions, newly added objects are instead written to a temporary
object directory and migrated to the primary object database upon
transaction commit.

When the --verbose option is specified, each individual object is
explicitly flushed via `flush_odb_transaction()` prior to reporting the
update. Flushing the object database transaction migrates pending
objects to the primary object database without marking the transaction
as complete. This is done so objects will be immediately visible to
git-update-index(1) callers using the --stdin option and that rely on
parsing verbose output to know when objects are written.

Using object database transactions in this manner is not really any
different than writing each object outside of a transaction.
Furthermore, the mechanism to flush and transaction without committing
is rather awkward. Drop the call to `flush_odb_transaction()` in favor
of skipping transactions altogether when git-update-index(1) has both
the --stdin and --verbose options specified.

Signed-off-by: Justin Tobler <jltobler@gmail.com>