bulk-checkin: drop transaction nesting
Object database transactions track a nested counter to ensure that only
a single transaction is active at a time. This leads to a rather awkward
interface as is becomes difficult to know if `begin_odb_transaction()`
or `end_odb_transaction()` are actually starting/completing a
transaction.

Instead of relying on the transaction itself to track nested
invocations, update transaction call sites that may be potentially
nested to check if there is already a pending transaction inflight.

Signed-off-by: Justin Tobler <jltobler@gmail.com>