diff: add option to report whether files are binary
When generating diff output, if either side of a filepair is detected as
binary, Git omits the diff content and instead prints a "Binary files
differ" message.

In some cases, it can be useful to know which files were detected as
binary and which were treated as text. Introduce a new
`--report-binary-files` diff option that, when enabled, appends a line
for each file indicating whether Git considered it binary or text.

Signed-off-by: Justin Tobler <jltobler@gmail.com>